Loot & Drops
Enemies now drop loot, and each can have their own loot table or several, with a setting for chances for a mix of guaranteed and rare loot tables.

Will need to add an in-memory item database, and ideally item id generation using guids. Also need to implement an inventory to keep track of all collected items.